requirements.txt
jishaku @ git+git://github.com/Gorialis/jishaku.git@873f863052a663efab3b6184d2516e03bf16bcf9
discord-ext-menus @ git+git://github.com/Rapptz/discord-ext-menus.git@309c70222eec33acacb8c21bbb0fbe9bc57de661
discord-ext-wizards==0.3.0
discord-ext-prettyhelp==0.2.2
requests==2.25.1
discord.py==1.7.2
humanize==3.5.0
quart==0.15.0
alt-discord-flags==0.0.5
discord-py-slash-command==1.1.2
python-dotenv==0.17.1
websockets==9.0.2
uvloop==0.15.2
asyncpg==0.23.0
buildpg==0.2.3
aiocache==0.11.1
ujson==4.0.2
msgpack==1.0.2
emoji==1.2.0
patreon==0.5.0
quart-csrf==0.2
quart-discord==2.1.4
